+++ src/App.jsx
@@ -5,7 +5,7 @@ import Footer from './components/footer/Footer.jsx';
 import Main from './components/main/Main.jsx';
 import Header from './components/header/Header.jsx';
 import Favorites from './components/main/Favorites.jsx';
-import { FavoritesProvider } from './hooks/FavoritesContext.jsx';
+import { FavoritesProvider } from './contexts/FavoritesContext.jsx';
 import SearchResults from './components/main/SearchResults.jsx';
 
 import './App.css';
