+++ src/components/header/Header.jsx
@@ -1,18 +1,22 @@
 import face from '../../assets/faceofmasculinity.jpg';
 import Categories from './Categories.jsx';
 import Search from './SearchField.jsx';
-import { Link } from 'react-router-dom';
+import { Link, useLocation } from 'react-router-dom';
 
 const Header = ({ chooseCategory }) => {
+  const location = useLocation();
+  const isFavoritesPage = location.pathname === '/favorites';
+  const isSearchResultsPage = location.pathname === '/search';
+
   return (
     <>
       <nav>
         <Link to="/">Main page</Link>
-        <Categories chooseCategory={chooseCategory} />
+        {!isFavoritesPage && !isSearchResultsPage && <Categories chooseCategory={chooseCategory} />}
         <Link to="/favorites">Favorites</Link>
       </nav>
 
-      <Search />
+      {!isFavoritesPage && <Search />}
 
       <h2>Chuck Norris facts generator</h2>
