+++ src/contexts/FavoritesContext.jsx
